requirements-core.txt
# Core runtime
pydantic>=2.8.0
httpx>=0.27.0
typer>=0.12.3
rich>=13.7.1
crawl4ai>=0.4.0,<1.0
beautifulsoup4>=4.12.3
lxml>=5.3.0
markdownify>=0.13.1
PyMuPDF>=1.26.0
pymupdf4llm>=0.0.27
pycryptodome>=3.20.0
# Validator runtime
websockets>=16.0
eth-account>=0.13.0
